cloud_io: adopt cloud_storage::auth_refresh_bg_op
