+++ src/v/cloud_io/CMakeLists.txt
@@ -1,6 +1,7 @@
 v_cc_library(
   NAME cloud_io
   SRCS
+    auth_refresh_bg_op.cc
     io_resources.cc
   DEPS
     Seastar::seastar
